Rescue seal implemented
Still needs rescueminigame finalization. The seals get a random number that signifies the specific injury they have when they spawn, which we will use to determine the RescueMinigame.

public int injury;//1-Net entanglement,2- Hook in mouth, 3- Desinfect, 4- normal.

If for instance the seal spawns with 4 we just use the petting function finan created, but if we get the net injury, we have to do the "remove the net" mini game".