Scrollable scene, seal teleportation
Instead of working with the UI I decided to work with game objects since the seal is a game object and parenting it to the ui doesnt work. So I decided to make the camera moveable. Using the camera the player can "scroll" the habitat scene.
The seals now teleport to a vector 2 ( which has to be adjusted by going in the seal prefab) depending on health. I dont have the scene images so I didnt add any colliders to separate the different zones but this is very easy to do so have fun with it.